Time: 47 ms (44.93%), Space: 89.4 MB (43.34%) - LeetHub

+++ 4-median-of-two-sorted-arrays/4-median-of-two-sorted-arrays.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     double findMedianSortedArrays(vector<int>& nums1, vector<int>& nums2) {
         
-        float temp[nums1.size()+nums2.size()];
+        double temp[nums1.size()+nums2.size()];
         
         for(int i=0;i<nums1.size();i++)
             temp[i]=nums1[i];
